Editors now handled independant of Choo.
This is necessary as an integration into choo was causing redraws and
massive problems with focus. Mainly the loss of it.
We also introduce here a lot of effects. This gives even greater controll
and enables to have fewer reducers which we compose using actions to
achieve more complex actions in a controllable way.